[Add] #46 投稿詳細画面からのアイテム登録の実装
一度でも検索されたアイテムは登録できなかったためコードの追加と変更を行った。
