+++ app/controllers/items_controller.rb
@@ -7,30 +7,17 @@ class ItemsController < ApplicationController
     @search = params[:keyword]
     if @search
       @results = RakutenWebService::Ichiba::Product.search(keyword: params[:keyword], hits: 5)
+      #@itemsに検索して取得したデータ入れる
       @results.each do |result|
         item = Item.new(read(result))
         @items << item
-        #Itemテーブルに同じアイテムがあれば保存しない処理を書く
+        #Itemテーブルに同じアイテムがあれば保存しない処理
         @items.each do |item|
           unless Item.all.exists?(name: item.name)
             item.save
           end
         end
       end
-      #もし同じアイテムが検索されたらそのアイテムのidを取得する必要がある
-    end
-  end
-
-  def create
-    @item = Item.find(params[:format])
-    @post_item = PostItem.new(post_id: @post.id, item_id: @item.id)
-    binding.pry
-    if @post_item.save
-      binding.pry
-      redirect_to post_path(@post.id)
-      flash[:notice] = "アイテムを登録しました。"
-    else
-      flash[:alert] = "アイテムの登録に失敗しました."
     end
   end
 
@@ -50,9 +37,5 @@ class ItemsController < ApplicationController
 
   def set_post
     @post = Post.find_by(id: params[:post_id])
-  end
-
-  def create_params
-    params.permit(:item)
   end
 end
