+++ app/views/items/_item_list.html.erb
@@ -1,12 +1,17 @@
-
-<i class="fa-solid fa-plus"><%= link_to search_items_post_items_path(@post) do %>
-  アイテムを追加する
-<% end %></i>
-<div class="post_item">
-  <% @items.each do |item| %>
-    <td><%= image_tag(item.image_url) %></td>
-    <td><%= item.name%></td>
-    <td><%= item.price%></td>
+<div class="row">
+  <% if @items.present? %>
+    <% @items.each do |item| %>
+      <div class="col-md-2">
+        <div class="card" style="width: 6rem; height: 6rem;">
+          <%= image_tag(item.image_url) %>
+        </div>
+        <div class="card-body">
+          <%= item.name%>
+          <%= item.price%>
+        </div>
+      </div>
+    <% end %>
+  <% else %>
+    <p>この投稿にはアイテムが登録されていません</p>
   <% end %>
 </div>
-
