+++ app/views/items/search_items.html.erb
@@ -2,14 +2,14 @@
 <div class="wrapper">
   <p class="alert"><%= alert %></p>
   <div class="search">
-    <%= form_with url: search_items_post_items_path, method: :get, local: true do |f|%>
+    <%= form_with url: post_search_items_path, method: :get, local: true do |f|%>
       <%= f.text_field :keyword, required: true %>
       <%= f.submit '検索する', disable_with: '検索中...' %>
     <% end %>
   </div>
 
   <h2>検索結果</h2>
-  <div id='item_list'>
+  <div class='item_list'>
     <table class="table table-hover">
       <thead>
           <tr>
@@ -21,12 +21,11 @@
       </thead>
       <tbody>
         <% @items.each do |item| %>
-          <tr class="checkbox_row" id="<%=item.id%>">
+          <tr class="checkbox_row">
             <td><%= image_tag(item.image_url) %></td>
             <td><%= item.name%></td>
             <td><%= item.price%></td>
-            <td><%= item.id%></td>
-            <td><%= link_to 'アイテムを追加', post_items_path(@post.id, item.id), method: :post %></td>
+            <td><%= link_to 'アイテムを追加', post_items_path(post_id: @post.id, item_name: item.name), method: :post %></td>
           </tr>
         <% end %>
       </tbody>
